pyproject.toml
[project]
name = "challenge-mle-danielospina"
version = "0.1.0"
description = ""
authors = [
    {name = "DanielOspina2301",email = "daniel.alberto.ospina@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
[tool.poetry.dependencies]
fastapi = ">=0.86.0,<0.87.0"
pydantic = ">=1.10.2,<1.11.0"
uvicorn = ">=0.15.0,<0.16.0"
numpy = ">=1.22.4,<1.23.0"
pandas = ">=1.3.5,<1.4.0"
scikit-learn = ">=1.3.0,<1.4.0"
matplotlib = "3.8.0"
seaborn = "0.12.2"
xgboost = ">=2.1.2,<2.2.0"
google-cloud-storage = ">=2.18.2,<2.19.0"
google-cloud-bigquery = ">=3.26.0,<3.27.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
